requirements.txt
Django==5.0.4
djangorestframework==3.15.1
gspread==6.1.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-decouple==3.8
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
